requirements.txt
pyyaml
networkx
numpy
pandas
matplotlib
torch
torch-geometric
scikit-learn